+++ src/core/ModuleController.js
@@ -1,0 +1,20 @@
+class ModuleController {
+    constructor(virtualTerminal) {
+        this.terminal = virtualTerminal;
+        this.active = false;
+    }
+
+    activate() {
+        this.active = true;
+    }
+
+    deactivate() {
+        this.active = false;
+    }
+
+    handleInput(event) {
+        // Override in child classes
+    }
+}
+
+export default ModuleController;
